New testing grounds
Added a new file called test_function for testing...functions. The
launcher in RayStation is now set to point at one of the functions
there, depending on who is using the software. This allows DM to have
her own test functions in place even if I'm working on mine.
