+++ hmrlib.log
@@ -16,3 +16,15 @@
 2016-11-28 15:51:48,571 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
 2016-11-28 15:51:48,696 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] End of script testMA_NePasUtiliser.py [execution time = 0:00:05.912000]
 2016-11-28 15:51:48,817 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:03:49,217 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:03:49,653 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] Starting script testMA_NePasUtiliser.py
+2016-11-28 17:03:49,778 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:03:49,903 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:03:50,026 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] End of script testMA_NePasUtiliser.py [execution time = 0:00:00.858000]
+2016-11-28 17:03:50,148 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:04:37,603 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:04:37,847 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] Starting script testMA_NePasUtiliser.py
+2016-11-28 17:04:37,971 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:06:28,421 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
+2016-11-28 17:06:28,546 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] End of script testMA_NePasUtiliser.py [execution time = 0:01:50.871000]
+2016-11-28 17:06:28,667 - hmrlib - INFO - [hmr30489 - PICARD^ELISE (1208465) - CT 1 - Stereo Vertebre/Stereo] ================================================================================
